コントローラ操作(プレイヤー部分)はできたけど何故か武器が拾えない 画面切り替え部分はまだ

+++ ControllerInput.cpp
@@ -325,7 +325,7 @@ bool ControllerInput::IsPadButtonRelease(const size_t& button, const size_t& pad
 	return false;
 }
 
-int ControllerInput::IsPadStick(const size_t& stickAxis, const float& deadzone, const size_t& padNo)
+float ControllerInput::IsPadStick(const size_t& stickAxis, const float& deadzone, const size_t& padNo)
 {
 	if (padNo >= gamepad.size())
 	{
